+++ coe/coe-input.js
@@ -8,7 +8,7 @@
 
 module.exports = function(RED) {
     'use strict';
-    const { getUnitInfo, getDigitalStateKey, mergeNodeData, createEmptyState } = require('../lib/utils');
+    const { Validate, getUnitInfo, getDigitalStateKey, mergeNodeData, createEmptyState } = require('../lib/utils');
 
     // CoE Input Node (receiving values)
     function CoEInputNode(config) {
@@ -19,16 +19,17 @@ module.exports = function(RED) {
         node.cmiConfig = RED.nodes.getNode(config.cmiconfig);
     
         if (!node.cmiConfig) {
-            node.error("CMI Configuration missing");
+            node.error("No CMI config assigned to CoE Input node.");
+            node.status({fill:"red", shape:"ring", text:"coe-input.status.noconfig"});
             return;
         }
         
-        node.cmiAddress = node.cmiConfig.address;
+        node.cmiAddress = node.cmiConfig.address || "";
         node.coeVersion = node.cmiConfig.coeVersion || 1;
-        node.lang = node.cmiConfig.lang;
-        node.nodeNumber = parseInt(config.nodeNumber) || 0;
-        node.outputNumber = parseInt(config.outputNumber) || 1;
-        node.dataType = config.dataType || 'analog';
+        node.lang = node.cmiConfig?.lang || "en";
+        node.nodeNumber = Validate.node(config.nodeNumber, true);
+        node.outputNumber = Validate.output(config.outputNumber, node.coeVersion);
+        node.dataType = Validate.type(config.dataType);
    
         // State management for LKGVs (Last Known Good Values) per block
         node.canNodeState = {};
@@ -59,7 +60,7 @@ module.exports = function(RED) {
                 }
                 
                 // Filter dataType number
-                if (canNode.dataType !== node.dataType) {
+                if (canNode.dataType !== node.dataType.long) {
                     continue;
                 }
                 
@@ -81,7 +82,7 @@ module.exports = function(RED) {
                     continue;
                 }
 
-                if (node.dataType === 'analog') {
+                if (node.dataType.long === 'analog') {
                     value = output.value;
                     unit = output.unit;
                     state = value;
@@ -96,17 +97,17 @@ module.exports = function(RED) {
                 const unitInfo = getUnitInfo(unit, node.lang);
                 const msg = {
                     payload: value,
-                    topic: `coe/${node.nodeNumber || mergedCanNode.nodeNumber}/${node.dataType}/${node.outputNumber}`,
+                    topic: `coe/${node.nodeNumber || mergedCanNode.nodeNumber}/${node.dataType.long}/${node.outputNumber}`,
                     coe: {
-                        timestamp: received.timestamp,
                         sourceIP: received.sourceIP,
                         nodeNumber: mergedCanNode.nodeNumber,
-                        dataType: node.dataType,
+                        dataType: node.dataType.long,
                         outputNumber: node.outputNumber,
                         state: state,
                         unit: unit,
                         unitName: unitInfo.name,
-                        unitSymbol: unitInfo.symbol
+                        unitSymbol: unitInfo.symbol,
+                        timestamp: received.timestamp
                         // raw: received.rawBuffer ? received.rawBuffer.toString('hex').toUpperCase() : null
                     }
                 };
@@ -115,7 +116,7 @@ module.exports = function(RED) {
                      node.send(msg);
                 }
                 
-                currentNodeText = `${state} ${unitInfo.symbol || ''} [v${node.coeVersion}]` // Caching last Node text
+                currentNodeText = `${state} ${unitInfo.symbol || ''}` // Caching last Node text
 
                 node.status({
                     fill: "green", 
@@ -131,24 +132,34 @@ module.exports = function(RED) {
         };
 
         node.cmiConfig.registerListener(listener);
+
+        resetTimeout(); // Set initial timeout
         
         // Reset CoE Timeout
         function resetTimeout() {
             if (timeoutTimer) clearTimeout(timeoutTimer);
+
+            // Fallback status text (initial timeout)
+            const statusText = currentNodeText ? `${currentNodeText} (Timeout)` : node._("coe-input.status.initialTimeout");
+            
             timeoutTimer = setTimeout(() => {
-                node.status({ fill: "red", shape: "dot", text: `${currentNodeText} (Timeout)` });
+                node.status({ fill: "red", shape: "dot", text: statusText });
             }, timeoutMs);
         }
         
         // Status information, including if filtered
         if (node.nodeNumber === 0) {
-            node.status({fill: "yellow", shape: "ring", text: "coe-input.status.waitingAny"});
+            node.status({fill: "yellow", shape: "ring", text: node._("coe-input.status.waitingAny") + ` [v${node.coeVersion}]`});
         } else {
-            node.status({fill: "grey", shape: "ring", text: "coe-input.status.waiting"});
+            node.status({fill: "grey", shape: "ring", text: "ᐅ" + node.dataType.short + " " + node.nodeNumber + "/" + node.outputNumber + " " + node._("coe-input.status.waiting") + ` [v${node.coeVersion}]`});
         }
         
         node.on('close', function() {
             node.cmiConfig.unregisterListener(listener);
+
+            if (timeoutTimer) {
+                clearTimeout(timeoutTimer);
+            }
         });
 
     }