+++ coe/coe-output.js
@@ -8,7 +8,7 @@
 
 module.exports = function(RED) {
     'use strict';
-    const { getBlockInfo } = require('../lib/utils')
+    const { Validate, getBlockInfo } = require('../lib/utils')
     const { queueAndSend } = require('../lib/queueing');
 
     function CoEOutputNode(config) {
@@ -19,22 +19,24 @@ module.exports = function(RED) {
         node.cmiConfig = RED.nodes.getNode(config.cmiconfig);
 
         if (!node.cmiConfig) {
-            node.error("CoE Configuration missing or invalid.");
+            node.error("No CMI config assigned to CoE Output node.");
             node.status({fill:"red", shape:"ring", text:"coe-output.status.noconfig"});
             return;
         }
 
-        node.cmiAddress = node.cmiConfig.address;
+        node.cmiAddress = node.cmiConfig.address || "";
         node.coeVersion = node.cmiConfig.coeVersion || 1;
-        node.nodeNumber = parseInt(config.nodeNumber) || 1;
-        node.outputNumber = parseInt(config.outputNumber) || 1;
-        node.dataType = config.dataType || 'analog';
+        node.nodeNumber = Validate.node(config.nodeNumber, false);
+        node.outputNumber = Validate.output(config.outputNumber, node.coeVersion);
+        node.dataType = Validate.type(config.dataType);
         node.unit = parseInt(config.unit) || 0;
 
+        node.readyText = "ᐅ" + node.dataType.short + " " + node.nodeNumber + "/" + node.outputNumber + " " + node._("coe-output.status.ready") + ` [v${node.coeVersion}]`
+
         if (node.coeVersion === 2) {
-            node.queueKey = node.nodeNumber + '-' + node.dataType;
+            node.queueKey = node.nodeNumber + '-' + node.dataType.short;
         } else {
-            node.block = getBlockInfo(node.dataType, node.outputNumber);
+            node.block = getBlockInfo(node.dataType.short, node.outputNumber);
             node.queueKey = node.nodeNumber + '-' + node.block.number;
         }
         
@@ -52,18 +54,20 @@ module.exports = function(RED) {
         node.lastInputTime = 0;
         
         node.intervalTimer = null;
+
+        node.status({fill: "grey", shape: "ring", text: node.readyText});
     
         node.on('input', function(msg) {
             node.lastInputTime = Date.now();
             const timeSinceLastSend = (node.lastInputTime - node.lastSentTime) / 1000; // Seconds
-            const isSupressed = checkSuppressCondition(node, timeSinceLastSend);
-            node.lastReceivedMsg = msg;
-            
+
             // Prepare last received values
             const payloadValue = parseFloat(msg.payload);
             node.lastReceivedValue = isNaN(payloadValue) ? 0 : payloadValue;
             node.lastReceivedUnit = (msg.coe && msg.coe.unit !== undefined) ? parseInt(msg.coe.unit) : node.unit;
+            node.lastReceivedMsg = msg;
 
+            const isSupressed = checkSuppressCondition(node, timeSinceLastSend);
 
             // Filter based on sending conditions
             if (isSupressed > 0) { 
@@ -101,8 +105,6 @@ module.exports = function(RED) {
                 node.blockingTimer = null;
             }
         });
-        
-        node.status({fill:"grey", shape:"ring", text:node._("coe-output.status.ready") + ` [v${node.coeVersion}]`});
 
         // Start the sending interval timer
         function setIntervalTimer() {
@@ -129,7 +131,7 @@ module.exports = function(RED) {
                     });
                 }     
                 setTimeout(() => {
-                    node.status({fill: "grey", shape: "ring", text: node._("coe-output.status.ready") + ` [v${node.coeVersion}]`});
+                    node.status({fill: "grey", shape: "ring", text: node.readyText});
                 }, 5000);
                 
                 setIntervalTimer(); // Restart interval timer
@@ -179,9 +181,9 @@ module.exports = function(RED) {
             }
 
             // Check size of value change
-            if (node.dataType === 'analog') {
+            if (node.dataType.long === 'analog' && node.minChange > 0) {
                 const delta = Math.abs(node.lastReceivedValue - node.lastSentValue);
-                if (node.minChange > 0 && delta < node.minChange) {
+                if (delta < node.minChange) {
                     isSuppressed = 2;
                     blockReason = `Δ${delta.toFixed(2)} < ${node.minChange}`;
                 }